+++ src/vs/workbench/contrib/positronNotebook/browser/contrib/find/PositronFindWidget.css
